feat: move `sortMethodFuncs` to module `Sort`
